maybe a refactor is needed
